Move inner classes up. Use new CodeNarc rules.
